Time: 30 ms (90.65%), Space: 55 MB (86.99%) - LeetHub

+++ 0310-minimum-height-trees/0310-minimum-height-trees.java
@@ -25,7 +25,7 @@ class Solution {
             List<Integer> newLeaves = new ArrayList();
             for (int i : leaves) {
                 int near = adjList.get(i).iterator().next();
-                adjList.get(near).remove((Integer) i);
+                adjList.get(near).remove(i);
    
                 if (adjList.get(near).size() == 1) newLeaves.add(near);
             }
